+++ logstash_docker/start_logstash.sh
@@ -16,6 +16,10 @@ function updateFromEnv(){
     fi
 }
 
+#Change Sniffer IP to host's IP
+SNIF_IP=$(hostname -i)
+updateFromEnv "$SNIF_IP" "Sniffer IP address" $UDS_ETC/SniffersConfig.json "127.0.0.1" "$SNIF_IP"
+
 #Change log4j log level if needed
 if [[ "$UC_LOG_LEVEL" =~ ^(ALL|DEBUG|INFO|WARN|ERROR|FATAL|OFF|TRACE)$ ]]; then
     updateFromEnv "$UC_LOG_LEVEL" "UC_LOG_LEVEL" $UDS_ETC/log4j.properties "ERROR" "$UC_LOG_LEVEL"